Updates to support python virtualenvs.
These changes add a dependency on the python module if you
choose virtualenv as the provider.  The other noticable
change to the existing module is moving the beaver.init
file to be a template in support of virtualenv settings.

Also,
added the exclude attribute to beaver stanza.
add queue_timeout attribute to beaver.conf.